refactor(dot11): extract Dot11 layer into layer module
